property 2 @Value, MapStruct mappers, tak delete + refac

+++ application/src/main/java/online/ronikier/todo/Application.java
@@ -4,6 +4,7 @@ import lombok.extern.slf4j.Slf4j;
 import online.ronikier.todo.domain.Task;
 import online.ronikier.todo.infrastructure.TaskRepository;
 import online.ronikier.todo.library.Utilities;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -15,6 +16,9 @@ import org.springframework.data.neo4j.repository.config.EnableNeo4jRepositories;
 @EnableNeo4jRepositories
 public class Application {
 
+    @Value("${todo.setup.initialize.database:false}")
+    private boolean setupInitializeDatabase;
+
     public static void main(String[] args) {
         SpringApplication.run(Application.class, args);
     }
@@ -23,10 +27,13 @@ public class Application {
     CommandLineRunner initialize(TaskRepository taskRepository) {
         return args -> {
 
-            taskRepository.deleteAll();
-            Task millionDollarTask = new Task(true, true, Utilities.dateCurrent(), Utilities.dateCurrent(), Utilities.dateFuture(7), "Be Rich", "Million Dollars");
-            taskRepository.save(millionDollarTask);
+            if (setupInitializeDatabase) {
+                log.info(Messages.INFO_INITIALISING_DATABASE);
+                taskRepository.deleteAll();
+                Task millionDollarTask = new Task(true, true, Utilities.dateCurrent(), Utilities.dateCurrent(), Utilities.dateFuture(7), "Be Rich", "Million Dollars");
+                taskRepository.save(millionDollarTask);
 
+            }
         };
     }
